Delete Mono/RuntimeMarshal.cs
The methods in this file were partially redundant and partially specific to AssemblyName.
